requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.34.150
botocore==1.34.150
certifi==2024.7.4
charset-normalizer==3.3.2
dj-static==0.0.6
Django==5.0.7
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-storages==1.14.4
django-tinymce==4.1.0
idna==3.7
jmespath==1.0.1
lxml==5.3.0
python-dateutil==2.9.0.post0
razorpay==1.4.2
requests==2.32.3
s3transfer==0.10.2
setuptools==72.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
static-ranges==0.2.0
static3==0.7.0
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
